refactor: Request 객체 내 로직 삭제 및 입력 파싱 리팩터링
