+++ src/main/java/racingcar/dto/CarNamesRequest.java
@@ -1,23 +1,16 @@
 package racingcar.dto;
 
-import java.util.Arrays;
 import java.util.List;
-import java.util.stream.Collectors;
 
 public class CarNamesRequest {
 	private final List<String> carNames;
 
-	public static CarNamesRequest from(String carNames) {
+	public static CarNamesRequest from(List<String> carNames) {
 		return new CarNamesRequest(carNames);
 	}
 
-	private CarNamesRequest(String carNames) {
-		this.carNames = splitByComma(carNames);
-	}
-
-	private List<String> splitByComma(String carNames) {
-		String[] splitCarNames = carNames.split("\\s*,\\s*");
-		return Arrays.stream(splitCarNames).collect(Collectors.toList());
+	private CarNamesRequest(List<String> carNames) {
+		this.carNames = carNames;
 	}
 
 	public List<String> getCarNames() {
